+++ NEWS.md
@@ -2,9 +2,12 @@
 
 * overhauls the API (#15, with @AliciaSchep)
 
+  * changes `spec_version()` -> `vw_spec_version()`
+  * changes `examine()` -> `vw_examine()`
   * changes `block_retrieve()` -> `vw_retrieve_block()`
   * changes `block_build_directory()` -> `vw_block_build_directory()`
   * changes `block_index()` -> `vw_block_index()`
+  * changes `block_create_gistid()` -> `vw_create_block_gistid()`
   * changes `block_create()` -> `vw_create_block()`
   * changes `block_config()` -> `vw_block_config()`  
   * changes `autosize()` -> `vw_autosize()`